+++ Adj/Program/Datastructure/Property.hs
@@ -1,1 +1,3 @@
-module Adj.Program.Datastructure.Property () where
+module Adj.Program.Datastructure.Property (module Exports) where
+
+import Adj.Program.Datastructure.Property.Cardinality as Exports